refactor: Items refactor.
